% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{add_issues<-}
\alias{add_issues<-}
\title{adding issues together}
\usage{
add_issues(x) <- value
}
\description{
adding issues together
}
\examples{
# use me like you'd use `names()`:
add_issues(existing_list_of_issues)<-another_list_of_issues
# is equivalent to:
existing_list_of_issues<-rbind(existing_df_with_issues,another_df_with_issues)
}
