% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{check_input}
\alias{check_input}
\title{Check if Assessment files are in order}
\usage{
check_input(data, questionnaire, samplingframe, analysisplan)
}
\arguments{
\item{data}{the dataset ( from load_data() )}

\item{questionnaire}{the questionnaire ( from load_questionnaire() )}

\item{samplingframe}{the samplingframe ( from load_samplingframe() )}
}
\description{
Check if Assessment files are in order
}
