% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_input_analysisplan.R
\name{check_input_analysisplan}
\alias{check_input_analysisplan}
\title{check analysisplan
(this one is for analysisplan inputs to the hypegrammaR package specifically)}
\usage{
check_input_analysisplan(analysisplan)
}
\arguments{
\item{analysisplan}{the analysisplan as a data.frame}
}
\value{
list of issues (see ?new_issues() for details)
}
\description{
check analysisplan
(this one is for analysisplan inputs to the hypegrammaR package specifically)
}
