% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_input_data.R
\name{check_input_data}
\alias{check_input_data}
\title{Check if data format is good}
\usage{
check_input_data(data)
}
\arguments{
\item{data}{data as data frame}
}
\value{
data frame with issues (see `new_issues()` for details)
}
\description{
Check if data format is good
}
