% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_input_match_data_questionnaire.R
\name{check_input_match_data_questionnaire}
\alias{check_input_match_data_questionnaire}
\title{Check for inconsistencies between an xlsform questionnaire and a dataset}
\usage{
check_input_match_data_questionnaire(data, questions, choices)
}
\arguments{
\item{data}{dataset as data.frame}

\item{xlsform}{questions sheet as data.frame}

\item{xlsform}{choices sheet as data.frame}
}
\description{
Check for inconsistencies between an xlsform questionnaire and a dataset
}
