% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_input_questionnaire.R
\name{check_input_questionnaire}
\alias{check_input_questionnaire}
\title{Check if questionnaire format is good}
\usage{
check_input_questionnaire(questions, choices)
}
\arguments{
\item{questions}{kobo questions sheet as data frame}

\item{choices}{kobo choices sheet as data frame}
}
\value{
data frame with issues (see `new_issues()` for details)
}
\description{
Check if questionnaire format is good
}
