% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_input_samplingframe.R
\name{check_input_samplingframe}
\alias{check_input_samplingframe}
\title{check sampling frame data structure}
\usage{
check_input_samplingframe(samplingframe)
}
\value{
list of issues (see ?new_issues() for details)
}
\description{
check sampling frame data structure
}
