% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{new_issues}
\alias{new_issues}
\title{create a df with new issues in standard format}
\usage{
new_issues(issue = character(0),
  affected_files = character(length(issue)),
  affected_variables = character(length(issue)),
  severity = factor(length(issue), levels = c("minor", "problematic",
  "critical")), comment = character(length(issue)))
}
\description{
all parameters except 'issue' are optional
each paramter must be a vector, they all must have the same length length
'severity' must be one of the predefined levels
}
