% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_input.R
\name{check_input}
\alias{check_input}
\title{Check if Assessment files are in order}
\usage{
check_input(data = NULL, questions = NULL, choices = NULL,
  samplingframe = NULL, analysisplan = NULL)
}
\arguments{
\item{data}{the dataset ( from load_data() )}

\item{samplingframe}{the samplingframe ( from load_samplingframe() )}

\item{questionnaire}{the questionnaire ( from load_questionnaire() )}
}
\description{
Check if Assessment files are in order
}
